Reorder program pages
The biggest change to each was just moving the course sequence and transfer information to the bottom of each page. I mostly removed the Why Study sections, keeping some of the language on some of the pages. Also I added links to the academic catalog pages for each credential.

Someday I'll learn to commit early commit often.